refactor: 동네 설정 모달 네이밍 수정 (#17)
